feat: done multi node impl

+++ include/qubo_matrix.h
@@ -2,6 +2,6 @@
 
 #include <vector>
 
-std::vector<float> build_teacher_qubo_matrix_host(
-    const std::vector<float>& values, const std::vector<float>& weights,
-    float capacity, float P);
+std::vector<double> build_teacher_qubo_matrix_host(
+    const std::vector<double>& values, const std::vector<double>& weights,
+    double capacity, double P);
